Application configuration store
Like described in the `/apps` directory documentation [1] of the
"tmpl-go" template repository, "wand" also aims to support the
monorepo [2] layout.
In order to manage multiple applications, their information and metadata
are recorded in a configuration store where each entry is identified by
a unique ID, usually the name of the application. The `pkg/app` package
provides two interfaces and an unexported struct that implements it that
can be used through the exported `NewStore() Store` function.

- <T> `pkg/app.Config` — A `struct` type that holds information and
  metadata of an application.
- <I> `pkg/app.Store` — A storage that provides methods to record
  application configurations:
  - `Add(*Config)` — Adds a application configuration.
  - `Get(string) (*Config, error)` — Returns the application
    configuration for the given name or nil along with an error when not
    stored.
- <T> `appStore` — A storage for application configurations.
- <F> `NewStore() Store` — Creates a new store for application configurations.

[1]: https://github.com/svengreb/tmpl-go/tree/main/apps
[2]: https://trunkbaseddevelopment.com/monorepos

GH-8