Simplify project version switch/case block and error handling
Instead of surrounding the "switch/case" block, that checkt the
repository version based on the underlying VCS, with a "if" block to
handle the case where no repository type is set, it no simply includes
the "none" VCS kind as a noop case.

Also the returned error now simply creates a new error instead of
formatting the error message as new error.

GH-10