Remove redundant GitHub Action workflow job for tests in Windows (#13)
There are currently no OS specific tests so the `test` workflow job
now only runs on Linux. The additional run on "Windows" was unnecessary
and only wasted build minutes because each minute is calculated 2x when
running on "Windows". See the section about "Minute multipliers" in the
official GitHub Action Pricing documentation [1].

Next to this it also helps to improve the performance since workflows
running on "Windows" are currently taking the most time.

[1]: https://docs.github.com/en/free-pro-team@latest/github/setting-up-and-managing-billing-and-payments-on-github/about-billing-for-github-actions

Closes GH-12