Go toolchain "caster"
To use the Go toolchain, also known as the `go` command [1], a new
caster [2], introduced in GH-14 [3], has been implemented.
To unify the handling of errors in the cast [5] package, a new
`ErrCast` [4] struct type has also been implemented.

The `Validate` function [6] of the new caster returns an error of type
`*cast.ErrCast` when the `go` binary executable does not exists at the
configured path or when it is also not available in the executable
search paths of the current environment.

[1]: https://golang.org/cmd/go
[2]: https://pkg.go.dev/github.com/svengreb/wand/pkg/cast#Caster
[3]: https://github.com/svengreb/wand/issues/14
[4]: https://pkg.go.dev/github.com/svengreb/wand/pkg/cast#ErrCast
[5]: https://pkg.go.dev/github.com/svengreb/wand/pkg/cast
[6]: https://pkg.go.dev/github.com/svengreb/wand/pkg/cast#Validate

GH-20