Task for Go toolchain `install` command
As of Go version 1.16 `go install $pkg@$version` [1] allows to install
commands without affecting the `main` module. Additionally commands
like `go build` and `go test` no longer modify `go.mod` and `go.sum`
files by default but report an error if a module requirement or checksum
needs to be added or updated (as if the `-mod=readonly` flag were used).
This can be used as alternative to the already existing `gobin` runner
[2].

To support the `go install` command of the Go toolchain [3], a new
`Task` [4] has been implemented in the new `install` [5] package that
can be used through a Go toolchain `Runner` [6].
The task is customizable through the following functions:

- `WithEnv(env map[string]string) install.Option` - sets the task
  specific environment.
- `WithModulePath(path string) install.Option` - sets the module import
  path.
- `WithModuleVersion(version *semver.Version) install.Option` - sets the
  module version.

[1]: https://blog.golang.org/go116-module-changes#TOC_4.
[2]: https://pkg.go.dev/github.com/svengreb/wand@v0.4.1/pkg/task/gobin
[3]: https://pkg.go.dev/cmd/go#hdr-Compile_and_install_packages_and_dependencies
[4]: https://pkg.go.dev/github.com/svengreb/wand/pkg/task#Task
[5]: https://pkg.go.dev/github.com/svengreb/wand/pkg/task/golang/install
[6]: https://pkg.go.dev/github.com/svengreb/wand/pkg/task/golang#Runner

GH-70