Remove `pkger` task in favor of Go 1.16 `embed` package
In #52 [1] a task for the github.com/markbates/pkger [2] Go module was
added, a tool for embedding static files into Go binaries.
The issue also includes the "Official Static Assets Embedding" section
which mentions that the task might be removed later on again as soon as
Go 1.16 [3] will be released which comes with toolchain support for
embedding static assets (files) [4] through the `embed` package [5].
Also see markbates/pkger#114 [6] for more details about the project
future of `pkger`.

The `pkger` package [7] has been removed and the `//go:embed` directive
should be used instead.

[1]: https://github.com/svengreb/wand/issues/52
[2]: https://pkg.go.dev/github.com/markbates/pkger
[3]: https://blog.golang.org/go1.16
[4]: https://golang.org/doc/go1.16#library-embed
[5]: https://pkg.go.dev/embed
[6]: https://github.com/markbates/pkger/issues/114
[7]: https://pkg.go.dev/github.com/svengreb/wand@v0.4.1/pkg/task/pkger

GH-74