Remove unnecessary `Wand` parameter in `Task` creation functions
Before most `Task` creation functions [^1] [^2] [^3] [^4] required a
`Wand` as parameter which was not used but blocked the internal usage
for task runners. Therefore these parameters have been removed.
When necessary, it can be added individually later on or can be
reintroduced through a dedicated function with extended parameters to
cover different use cases.

[^1]: https://pkg.go.dev/github.com/svengreb/wand@v0.5.0/pkg/task/gofumpt#New
[^2]: https://pkg.go.dev/github.com/svengreb/wand@v0.5.0/pkg/task/goimports#New
[^3]: https://pkg.go.dev/github.com/svengreb/wand@v0.5.0/pkg/task/golang/build#New
[^4]: https://pkg.go.dev/github.com/svengreb/wand@v0.5.0/pkg/task/golang/install#New

GH-76