`RunOut` method for `Runner` interface
The `Run` method of the `Runner` interface [1] allowes to run a command,
but does not return its output. This can be blocking when running
commands like `go env GOBIN` to get the path to the `GOBIN` environment
variable [2].

To support such uses cases, the `RunOut(Task) (string, error)`
method has been added to the `Runner` interface that runs a command and
returns its output.

[1]: https://pkg.go.dev/github.com/svengreb/wand@v0.5.0/pkg/task#Runner
[2]: https://pkg.go.dev/cmd/go#hdr-Print_Go_environment_information

GH-83