Remove unnecessary `app.Config` parameter from `Task` creation functions
Some functions that create a `Task` [1] required an `app.Config`
struct [2], but most tasks did not use the data in any way. To improve
the code quality and simplify the internal usage of tasks these
parameters have been removed as well as the field from the structs that
implement the `Task` interfaces.

[1]: https://pkg.go.dev/github.com/svengreb/wand/pkg/task#Task
[2]: https://pkg.go.dev/github.com/svengreb/wand@v0.5.0/pkg/app#Config

GH-85