Remove `ExecName` method from `Exec` interface again
In GH-79 [1] a new `ExecName() string` method has been added to the
`Exec` interface [2] to improve the way how `RunnerExec` [3] handles the
name of the executable file, but it turned out that this is not useful
at all and also broke the separation of concern between the `Task` and
`Runner` interfaces.
Therefore the method has been removed again.

[1]: https://github.com/svengreb/wand/issues/79
[2]: https://pkg.go.dev/github.com/svengreb/wand/pkg/task#Exec
[3]: https://pkg.go.dev/github.com/svengreb/wand/pkg/task#RunnerExec

GH-87